refactor: admin update appointment
