+++ src/app/api/v1/appointments/client/route.ts
@@ -40,7 +40,6 @@ export async function POST(request: Request) {
       return nextResponse;
     }
 
-    // 3. Book Appointment
     const user = await readCurrentUser();
     const newAppointment: Appointment = {
       appointmentDate: new Date(json.appointmentDate),
@@ -70,47 +69,8 @@ export async function PATCH() {
 export async function DELETE() {
   return NextResponse.json('Forbidden', { status: 403 });
 }
-export async function PUT(request: Request) {
-  const verifyUserHasRouteAccessHandler = new VerifyUserHasRouteAccessHandler();
-  const verifyAppointmentSchemaHandler = new VerifyAppointmentSchemaHandler();
-  const verifyBusinessIsOpenOnWeekdayHandler =
-    new VerifyBusinessIsOpenOnWeekdayHandler();
-  const verifyAppointmentIsBetweenOpeningHoursHandler =
-    new VerifyAppointmentIsBetweenOpeningHoursHandler();
-  const verifySellerIsAvailableHandler = new VerifySellerIsAvailableHandler();
-
-  verifyUserHasRouteAccessHandler
-    .setNext(verifyAppointmentSchemaHandler)
-    .setNext(verifyBusinessIsOpenOnWeekdayHandler)
-    .setNext(verifyAppointmentIsBetweenOpeningHoursHandler)
-    .setNext(verifySellerIsAvailableHandler);
-
-  try {
-    const json = await request.json();
-    const nextResponse = await verifyUserHasRouteAccessHandler.handle(json);
-
-    if (nextResponse !== null) {
-      return nextResponse;
-    }
-
-    const newAppointment: Appointment = {
-      appointmentDate: new Date(json.appointmentDate),
-      clientEmail: json.clientEmail,
-      clientName: json.clientName,
-      bookedAt: getUTCDate(new Date()),
-      sellerId: new ObjectId(json.sellerId),
-      clientNotes: json.clientNotes,
-    };
-
-    const result = await createAppointment(newAppointment);
-    if (!result) {
-      return NextResponse.json('Failed', { status: 400 });
-    }
-
-    return NextResponse.json({ result }, { status: 200 });
-  } catch (error) {
-    return NextResponse.json('Forbidden1', { status: 403 });
-  }
+export async function PUT() {
+  return NextResponse.json('Forbidden1', { status: 403 });
 }
 export async function OPTIONS() {
   return NextResponse.json('Forbidden', { status: 403 });
