+++ src/lib/validation/appointment/form-appointment.ts
@@ -3,10 +3,10 @@ import { z } from 'zod';
 /**
  * Validate the input client side
  */
-export const formAppointmentsSchema = z.object({
-  bookedAt: z.date(),
-  clientEmail: z.string(),
-  clientName: z.string(),
-  clientNotes: z.string(),
-  sellerId: z.any(),
+export const appointmentFormSchema = z.object({
+  bookingDate: z.date(),
+  bookingTimeSlotStart: z.string(),
+  sellerId: z.string(),
+  clientName: z.string().min(3),
+  clientEmail: z.string().optional(),
 });
